requirements.txt

streamlit
torch
torchvision
Pillow
timm
numpy
matplotlib
